consolidate subcommand definitions into subcommands.rs
Moving all the subcommand definitions into subcommands.rs lets us put
all of the documentation for the subcommands in a single place
